+++ app/[[...slug]]/metadata.ts
@@ -1,0 +1,6 @@
+import { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: "Mantine Next.js template",
+  description: "I am using Mantine with Next.js!",
+};