+++ next.config.mjs
@@ -1,3 +1,9 @@
-const nextConfig = {reactStrictMode: true,};
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  reactStrictMode: true,
+  experimental: {
+    optimizePackageImports: ['@mantine/core', '@mantine/hooks', '@mantine/notifications'],
+  },
+};
 
 export default nextConfig;