+++ src/app/components/styles/SelfHelpStyles.ts
@@ -1,9 +1,0 @@
-export { default as ButtonStyle } from './ButtonStyle';
-export { default as CardStyle } from './CardStyle';
-export { default as CarouselStyle } from './CarouselStyle';
-export { default as ContainerStyle } from './ContainerStyle';
-export { default as DivStyle } from './DivStyle';
-export { default as HeadingStyle } from './HeadingStyle';
-export { default as ImageStyle } from './ImageStyle';
-export { default as LinkStyle } from './LinkStyle';
-export { default as MarkdownStyle } from './MarkdownStyle';