+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 v0.0.2
- - replace MUI with Mantine
+ - replace `MUI` with `Mantine`
  - Mantine v7
+ - add `tailwind` css
+ - add `refine`
 
 v0.0.1
 - Initial commit